Update README.md
Just copying my notes from when I was going through this. Obviously all changes are entirely up to you, these are just suggestions (and a few mild typo fixes):

STRONGLY advise a revision of the line in the readme to "Adjust the download and installation paths." This is my least favorite line in so many of these readmes, and you will get endless questions about clarification for this.
Will also have to rephrase the step prior to this once official, since it will be on the UI under Browse Modlists.
Also strongly advise clarification on Game Folder Files. I know it looks fine and it isn't wrong, but...people are dumb, man. Everyone screws this up. I have a whole paragraph in the TPF readme just to be extremely clear about what they need to copy and paste.
You may also want to tell people not to ask about Ordinator unless you want to have an answer/profile ready for them. It's going to happen. A lot.
Hell, you might just want to steal the whole FAQ part of the TPF readme, honestly. People are always confused about QuickLight and iHud and whatnot.